drivers,include: use misc device to reuse device numbers
Previously, on each new found that was not
enumerated before, was allocated a /dev/fpga[0-11]
dev node. However, on removal, the respective device node
was not deallocated by the driver

Now, using misc devices, each new device is inserted
on a list, uniquely, and the respective device is
removed correctly.